# Support custom key type