--- a/integration_tests/serde/Cargo.toml
+++ b/integration_tests/serde/Cargo.toml
@@ -8,4 +8,5 @@ publish = false
 [dev-dependencies]
 intmap = { path = "../..", features = ["serde"] }
 proptest = "1.0.0"
+serde = "1.0.0"
 serde_json = "1.0.1"