--- a/src/entry.rs
+++ b/src/entry.rs
@@ -1,21 +1,21 @@
 // ***************** Entry *********************
 
-use crate::IntMap;
+use crate::{int::SealedInt, IntKey, IntMap};
 
 /// A view into a single entry in a [`IntMap`], which may either be vacant or occupied.
 ///
 /// The entry can be constructed by calling [`IntMap::entry`] with a key. It allows inspection
 /// and in-place manipulation of its value without repeated lookups.
-pub enum Entry<'a, V: 'a> {
+pub enum Entry<'a, K: IntKey, V: 'a> {
     /// The entry is occupied.
-    Occupied(OccupiedEntry<'a, V>),
+    Occupied(OccupiedEntry<'a, K, V>),
     /// The entry is vacant.
-    Vacant(VacantEntry<'a, V>),
+    Vacant(VacantEntry<'a, K, V>),
 }
 
-impl<'a, V> Entry<'a, V> {
+impl<'a, K: IntKey, V> Entry<'a, K, V> {
     #[inline]
-    pub(crate) fn new(key: u64, int_map: &'a mut IntMap<V>) -> Self {
+    pub(crate) fn new(key: K, int_map: &'a mut IntMap<K, V>) -> Self {
         let indices = Self::indices(key, int_map);
 
         match indices {
@@ -28,33 +28,34 @@ impl<'a, V> Entry<'a, V> {
         }
     }
 
-    fn indices(key: u64, int_map: &IntMap<V>) -> Option<(usize, usize)> {
+    fn indices(key: K, int_map: &IntMap<K, V>) -> Option<(usize, usize)> {
         if int_map.is_empty() {
             return None;
         }
 
-        let cache_ix = int_map.calc_index(key);
+        let k = key.into_int();
+        let cache_ix = k.calc_index(int_map.mod_mask);
 
         let vals = &int_map.cache[cache_ix];
         let vals_ix = { vals.iter() }
             .enumerate()
-            .find_map(|(vals_ix, &(k, _))| (k == key).then(|| vals_ix))?;
+            .find_map(|(vals_ix, (key, _))| (key.into_int() == k).then(|| vals_ix))?;
 
         Some((cache_ix, vals_ix))
     }
 }
 
 /// A view into an occupied entry in a [`IntMap`]. It is part of the [`Entry`] enum.
-pub struct OccupiedEntry<'a, V: 'a> {
+pub struct OccupiedEntry<'a, K: IntKey, V: 'a> {
     // Index to vals, guaranteed to be valid
     vals_ix: usize,
     // Element of IntMap::cache, guaranteed to be non-empty
-    vals: &'a mut Vec<(u64, V)>,
+    vals: &'a mut Vec<(K, V)>,
     // IntMap::count, guaranteed to be non-zero
     count: &'a mut usize,
 }
 
-impl<'a, V> OccupiedEntry<'a, V> {
+impl<'a, K: IntKey, V> OccupiedEntry<'a, K, V> {
     /// Gets a reference to the value in the entry.
     pub fn get(&self) -> &V {
         // Safety: We didn't modify the cache since we calculated the index
@@ -90,12 +91,12 @@ impl<'a, V> OccupiedEntry<'a, V> {
 }
 
 /// A view into a vacant entry in a [`IntMap`]. It is part of the [`Entry`] enum.
-pub struct VacantEntry<'a, V: 'a> {
-    key: u64,
-    int_map: &'a mut IntMap<V>,
+pub struct VacantEntry<'a, K: IntKey, V: 'a> {
+    key: K,
+    int_map: &'a mut IntMap<K, V>,
 }
 
-impl<'a, V: 'a> VacantEntry<'a, V> {
+impl<'a, K: IntKey, V: 'a> VacantEntry<'a, K, V> {
     pub fn insert(self, value: V) -> &'a mut V {
         self.int_map.insert(self.key, value);
         return self.int_map.get_mut(self.key).unwrap();