--- a/src/serde.rs
+++ b/src/serde.rs
@@ -1,11 +1,15 @@
-use crate::IntMap;
+use crate::{IntKey, IntMap};
 use serde::{
     de::{Deserializer, MapAccess, Visitor},
     ser::SerializeMap,
     Deserialize, Serialize, Serializer,
 };
 
-impl<T: Serialize> Serialize for IntMap<T> {
+impl<K, V> Serialize for IntMap<K, V>
+where
+    K: IntKey + Serialize,
+    V: Serialize,
+{
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
         S: Serializer,
@@ -18,7 +22,11 @@ impl<T: Serialize> Serialize for IntMap<T> {
     }
 }
 
-impl<'de, T: Deserialize<'de>> Deserialize<'de> for IntMap<T> {
+impl<'de, K, V> Deserialize<'de> for IntMap<K, V>
+where
+    K: IntKey + Deserialize<'de>,
+    V: Deserialize<'de>,
+{
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
     where
         D: Deserializer<'de>,
@@ -27,11 +35,11 @@ impl<'de, T: Deserialize<'de>> Deserialize<'de> for IntMap<T> {
     }
 }
 
-struct IntMapVisitor<V> {
-    marker: std::marker::PhantomData<fn() -> IntMap<V>>,
+struct IntMapVisitor<K: IntKey, V> {
+    marker: std::marker::PhantomData<fn() -> IntMap<K, V>>,
 }
 
-impl<V> IntMapVisitor<V> {
+impl<K: IntKey, V> IntMapVisitor<K, V> {
     fn new() -> Self {
         IntMapVisitor {
             marker: std::marker::PhantomData,
@@ -39,11 +47,12 @@ impl<V> IntMapVisitor<V> {
     }
 }
 
-impl<'de, V> Visitor<'de> for IntMapVisitor<V>
+impl<'de, K, V> Visitor<'de> for IntMapVisitor<K, V>
 where
+    K: IntKey + Deserialize<'de>,
     V: Deserialize<'de>,
 {
-    type Value = IntMap<V>;
+    type Value = IntMap<K, V>;
 
     fn expecting(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "IntMap<{}>", std::any::type_name::<V>())