This PR is an alternative to:

- #63, my initial PR that tries to avoid breaking changes
- #67, a PR from @414owen that allows any types as keys if they implement enough traits from the `num-traits` crate

This PR is a combination of both PRs:

- It supports other primitive integers
- It supports user provided types as long as the type is based on a primitive integer
- It requires a breaking change
- It does not require any additional dependencies

To achieve that two new traits are introduced. The first trait is a sealed trait for primitive integers:

```rust
pub trait Int: SealedInt {}
```

It defines the properties of an integer. It's implemented for `u8`, `u16`, `u32`, `u64`, `u128`, `usize`, `i8`, `i16`, `i32`, `i64`, `i128` and `isize`.

The second trait is an open trait for user provided keys:

```rust
pub trait IntKey: Copy {
    type Int: Int;
    fn into_int(self) -> Self::Int;
}
```

It defines how the key can be mapped to an integer. It's implemented for:

- `u8`, `u16`, `u32`, `u64`, `u128`, `usize`, `i8`, `i16`, `i32`, `i64`, `i128` and `isize`
- `std::num::NonZeroU8`, `std::num::NonZeroU16`, ...
- ~`std::num::Saturating`~ I removed it because it's not supported by the current MSRV
- `std::num::Wrapping`
- `std::net::Ipv4Addr`
- `std::net::Ipv6Addr`

And it can be implemented by the user.

`IntMap` has now a type parameter `K`:

```rust
pub struct IntMap<K, V> {
    // ...
}

impl<K: IntKey, V> IntMap<K, V> {
    // ...
}
```

Other changes worth mentioning:

- I changed the prime number for `u64`, see #68
- Some iterators had unnecessary lifetime bounds on `K` or `V`, e.g. `V: 'a`
- Some benchmarks used `&u64` as key instead of `u64` because of type inference, I made it consistent

Closes #61
Closes #68