@jakoschiko 2024-11-01
`IntKey` could also provide the prime number:

```rust
pub trait IntKey: Copy {
    type Int: Int;
    const PRIME: Self::Int;
    fn into_int(self) -> Self::Int;
}
```