@jakoschiko 2024-11-01
Benchmarks before:

```
basic_bench                              fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ u64_get_intmap                        12.31 µs      │ 35.8 µs       │ 13.1 µs       │ 13.81 µs      │ 100     │ 100
├─ u64_insert_intmap                     28.6 µs       │ 160 µs        │ 30.17 µs      │ 32.15 µs      │ 100     │ 100
├─ u64_insert_intmap_checked             29.36 µs      │ 139 µs        │ 32.8 µs       │ 34.18 µs      │ 100     │ 100
├─ u64_insert_intmap_entry               53.7 µs       │ 153.5 µs      │ 60.21 µs      │ 61.21 µs      │ 100     │ 100
├─ u64_insert_without_capacity_intmap    480.1 µs      │ 614.6 µs      │ 489.3 µs      │ 507.2 µs      │ 100     │ 100
╰─ u64_resize_intmap                     22.28 µs      │ 29.86 µs      │ 22.32 µs      │ 22.55 µs      │ 100     │ 100
```

Benchmarks after:

```
basic_bench                              fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ u64_get_intmap                        11.61 µs      │ 36.21 µs      │ 11.68 µs      │ 12.42 µs      │ 100     │ 100
├─ u64_insert_intmap                     27.5 µs       │ 134.1 µs      │ 28.1 µs       │ 29.65 µs      │ 100     │ 100
├─ u64_insert_intmap_checked             35.06 µs      │ 143.9 µs      │ 41.44 µs      │ 42.34 µs      │ 100     │ 100
├─ u64_insert_intmap_entry               44.57 µs      │ 149.9 µs      │ 47.27 µs      │ 48.89 µs      │ 100     │ 100
├─ u64_insert_without_capacity_intmap    472.6 µs      │ 520.3 µs      │ 478.9 µs      │ 481.3 µs      │ 100     │ 100
╰─ u64_resize_intmap                     20.82 µs      │ 24.84 µs      │ 20.88 µs      │ 21.01 µs      │ 100     │ 100
```

I also added some new benchmarks for comparing different keys:

```
key_comparison              fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ get                                    │               │               │               │         │
│  ├─ u128                  20 µs         │ 40.59 µs      │ 20.17 µs      │ 20.84 µs      │ 100     │ 100
│  ├─ u16                   12.77 µs      │ 37.23 µs      │ 12.85 µs      │ 13.89 µs      │ 100     │ 100
│  ├─ u32                   13.19 µs      │ 35.06 µs      │ 13.59 µs      │ 14.16 µs      │ 100     │ 100
│  ╰─ u64                   12.71 µs      │ 36.19 µs      │ 13.53 µs      │ 14.19 µs      │ 100     │ 100
├─ insert                                 │               │               │               │         │
│  ├─ u128                  42.5 µs       │ 149.8 µs      │ 43.57 µs      │ 45.33 µs      │ 100     │ 100
│  ├─ u16                   34.1 µs       │ 141.7 µs      │ 35.19 µs      │ 37.71 µs      │ 100     │ 100
│  ├─ u32                   30.38 µs      │ 141.3 µs      │ 31.01 µs      │ 32.82 µs      │ 100     │ 100
│  ╰─ u64                   33.68 µs      │ 156.5 µs      │ 34.37 µs      │ 38.06 µs      │ 100     │ 100
╰─ insert_without_capacity                │               │               │               │         │
   ├─ u128                  508.8 µs      │ 590.1 µs      │ 520.5 µs      │ 522.3 µs      │ 100     │ 100
   ├─ u16                   483.9 µs      │ 617.7 µs      │ 489.2 µs      │ 492.3 µs      │ 100     │ 100
   ├─ u32                   487.8 µs      │ 515.6 µs      │ 494.4 µs      │ 495.2 µs      │ 100     │ 100
   ╰─ u64                   504.8 µs      │ 532.8 µs      │ 508.9 µs      │ 510.1 µs      │ 100     │ 100
```