@JesperAxelsson 2024-11-02
I like this one the best so far. Adding direct support for ip addresses were a nice touch :)

You have a few `let int = key.into_int();`, it might be worth considering shadowing key instead. As far as `get` or `insert` is concerned it's a key. Just a small nitpick.

I'll wait a few days to let @414owen come with feedback, then I'll merge and create a new major release.

Otherwise, great job and thank you for your patience.