@jakoschiko 2024-11-02
> Adding direct support for ip addresses were a nice touch :)

:p

> You have a few `let int = key.into_int();`, it might be worth considering shadowing key instead. As far as `get` or `insert` is concerned it's a key. Just a small nitpick.

This does not work in all cases because the actual key is used afterwards. I renamed `int` to `k`, a shorter name for a variable with a smaller scope.

> I'll wait a few days to let @414owen come with feedback, then I'll merge and create a new major release.

Sounds good. If you don't mind, I'd also want to complete #56 and #70 before the next major release.

What do you think about adding `const PRIME: Self::Int;` to the trait `IntKey`? Then a user could choose another prime number by using a wrapper for the key.