@JesperAxelsson 2024-11-07
> Sounds good. If you don't mind, I'd also want to complete #56 and #70 before the next major release.
> 
Sure, I don't mind waiting a bit.

> What do you think about adding `const PRIME: Self::Int;` to the trait `IntKey`? Then a user could choose another prime number by using a wrapper for the key.
Yeah, that might be useful if someone want to use 256 bit or 7 bit sized key for some reason.