@jakoschiko 2024-11-07
`Int` represents a primitive integer and the user can't implement it. 256 bit is not possible because there is no primitive integer with 256 bit. But I would assume that for such types `HashMap` is better anyway.