@jakoschiko 2024-11-07
I made some changes:

- `Int` is now implemented only for unsigned integers. That makes it easier to support custom prime numbers. Note that signed integers can still be used as key because they implement `IntKey`.
- I added support for custom prime numbers via `IntKey::Prime`.
- I improved the doc. `IntKey` has now an example that shows how it can be implemented.
- I removed unnecessary trait bounds for `Int`. This is just an internal cleanup.